server/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
python-dotenv>=1.0.1
pydantic>=2.10.0
psycopg2-binary>=2.9.10
redis>=5.2.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.3.0
chromadb>=0.5.0
jinja2>=3.1.4
reportlab>=4.2.0
requests>=2.32.0
python-multipart>=0.0.12
